split data modules by task
